requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
openai==1.40.3
google-generativeai==0.8.3
elevenlabs==1.9.0
sounddevice==0.4.7
soundfile==0.12.1
speechrecognition==3.10.4
pydub==0.25.1
